Functions to plot
The first function that gets plotted is the actual prime counting function, and then there are some under-approximations (that get pretty close it) in there. The fact that these are underapproximations shows that the function field PNT isn't really an exact analogy of the integer version prime counting function (though the leading terms agree).